Fix and upgrade Mcode
