+++ MK4duo/src/bedlevel/least_squares_fit.cpp
@@ -65,4 +65,4 @@
     return 0;
   }
 
-#endif // AUTO_BED_LEVELING_UBL || ENABLED(AUTO_BED_LEVELING_LINEAR)  
+#endif // ENABLED(AUTO_BED_LEVELING_LINEAR)
