+++ MK4duo/src/gcode/lcd/m0_m1.h
@@ -1,0 +1,95 @@
+/**
+ * MK4duo 3D Printer Firmware
+ *
+ * Based on Marlin, Sprinter and grbl
+ * Copyright (C) 2011 Camiel Gubbels / Erik van der Zalm
+ * Copyright (C) 2013 - 2017 Alberto Cotronei @MagoKimbra
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+
+/**
+ * mcode
+ *
+ * Copyright (C) 2017 Alberto Cotronei @MagoKimbra
+ */
+
+#if HAS_RESUME_CONTINUE
+
+  #define CODE_M0
+  #define CODE_M1
+
+  /**
+   * M0: Unconditional stop - Wait for user button press on LCD
+   * M1: Conditional stop   - Wait for user button press on LCD
+   */
+  inline void gcode_M0_M1(void) {
+    const char * const args = parser.string_arg;
+
+    millis_t ms = 0;
+    bool hasP = false, hasS = false;
+    if (parser.seenval('P')) {
+      ms = parser.value_millis(); // milliseconds to wait
+      hasP = ms > 0;
+    }
+    if (parser.seenval('S')) {
+      ms = parser.value_millis_from_seconds(); // seconds to wait
+      hasS = ms > 0;
+    }
+
+    #if ENABLED(ULTIPANEL)
+
+      if (!hasP && !hasS && *args != '\0')
+        lcd_setstatus(args, true);
+      else {
+        LCD_MESSAGEPGM(MSG_USERWAIT);
+        #if ENABLED(LCD_PROGRESS_BAR) && PROGRESS_MSG_EXPIRE > 0
+          dontExpireStatus();
+        #endif
+      }
+
+    #else
+
+      if (!hasP && !hasS && *args != '\0')
+        SERIAL_LV(ECHO, args);
+
+    #endif
+
+    printer.wait_for_user = true;
+    KEEPALIVE_STATE(PAUSED_FOR_USER);
+
+    stepper.synchronize();
+    commands.refresh_cmd_timeout();
+
+    if (ms > 0) {
+      ms += commands.previous_cmd_ms;  // wait until this time for a click
+      while (PENDING(millis(), ms) && printer.wait_for_user) printer.idle();
+    }
+    else {
+      #if ENABLED(ULTIPANEL)
+        if (lcd_detected()) {
+          while (printer.wait_for_user) printer.idle();
+          IS_SD_PRINTING ? LCD_MESSAGEPGM(MSG_RESUMING) : LCD_MESSAGEPGM(WELCOME_MSG);
+        }
+      #else
+        while (printer.wait_for_user) printer.idle();
+      #endif
+    }
+
+    printer.wait_for_user = false;
+    KEEPALIVE_STATE(IN_HANDLER);
+  }
+
+#endif // HAS_RESUME_CONTINUE
