+++ MK4duo/src/gcode/m_code.h
@@ -26,1010 +26,6 @@
  * Copyright (C) 2017 Alberto Cotronei @MagoKimbra
  */
 
-#if HAS_RESUME_CONTINUE
-
-  /**
-   * M0: Unconditional stop - Wait for user button press on LCD
-   * M1: Conditional stop   - Wait for user button press on LCD
-   */
-  inline void gcode_M0_M1() {
-    const char * const args = parser.string_arg;
-
-    millis_t ms = 0;
-    bool hasP = false, hasS = false;
-    if (parser.seenval('P')) {
-      ms = parser.value_millis(); // milliseconds to wait
-      hasP = ms > 0;
-    }
-    if (parser.seenval('S')) {
-      ms = parser.value_millis_from_seconds(); // seconds to wait
-      hasS = ms > 0;
-    }
-
-    #if ENABLED(ULTIPANEL)
-
-      if (!hasP && !hasS && *args != '\0')
-        lcd_setstatus(args, true);
-      else {
-        LCD_MESSAGEPGM(MSG_USERWAIT);
-        #if ENABLED(LCD_PROGRESS_BAR) && PROGRESS_MSG_EXPIRE > 0
-          dontExpireStatus();
-        #endif
-      }
-
-    #else
-
-      if (!hasP && !hasS && *args != '\0')
-        SERIAL_LV(ECHO, args);
-
-    #endif
-
-    printer.wait_for_user = true;
-    KEEPALIVE_STATE(PAUSED_FOR_USER);
-
-    stepper.synchronize();
-    commands.refresh_cmd_timeout();
-
-    if (ms > 0) {
-      ms += commands.previous_cmd_ms;  // wait until this time for a click
-      while (PENDING(millis(), ms) && printer.wait_for_user) printer.idle();
-    }
-    else {
-      #if ENABLED(ULTIPANEL)
-        if (lcd_detected()) {
-          while (printer.wait_for_user) printer.idle();
-          IS_SD_PRINTING ? LCD_MESSAGEPGM(MSG_RESUMING) : LCD_MESSAGEPGM(WELCOME_MSG);
-        }
-      #else
-        while (printer.wait_for_user) printer.idle();
-      #endif
-    }
-
-    printer.wait_for_user = false;
-    KEEPALIVE_STATE(IN_HANDLER);
-  }
-
-#endif // HAS_RESUME_CONTINUE
-
-#if HAS_MULTI_MODE
-
-  /**
-   * M3: Setting laser beam or fire laser - CNC clockwise speed
-   * M4: Turn on laser beam or CNC counter clockwise speed
-   *      S - Laser intensity or CNC speed
-   *      L - Laser Duration
-   *      P - PPM
-   *      D - Diagnostic
-   *      B - Set mode
-   */
-  inline void gcode_M3_M4(bool clockwise) {
-    stepper.synchronize();
-
-    switch (printer.mode) {
-
-      #if ENABLED(LASER) && ENABLED(LASER_FIRE_SPINDLE)
-        case PRINTER_MODE_LASER: {
-          if (IsRunning()) {
-            #if ENABLED(INTENSITY_IN_BYTE)
-              if (parser.seenval('S')) laser.intensity = (float)(parser.value_byte() / 255) * 100.0;
-            #else
-              if (parser.seenval('S')) laser.intensity = parser.value_float();
-            #endif
-            if (parser.seenval('L')) laser.duration = parser.value_ulong();
-            if (parser.seenval('P')) laser.ppm = parser.value_float();
-            if (parser.seenval('D')) laser.diagnostics = parser.value_bool();
-            if (parser.seenval('B')) laser.set_mode(parser.value_int());
-          }
-          laser.status = LASER_ON;
-        }
-        break;
-      #endif
-
-      #if ENABLED(CNCROUTER)
-        case PRINTER_MODE_CNC:
-          if (parser.seenval('S')) setCNCRouterSpeed(parser.value_ulong(), clockwise);
-        break;
-      #endif
-
-      default: break; // other tools
-
-    } // printer.mode
-
-    mechanics.prepare_move_to_destination();
-  }
-
-  /**
-   * M5: Turn off laser beam - CNC off
-   */
-  inline void gcode_M5() {
-    stepper.synchronize();
-
-    switch (printer.mode) {
-
-      #if ENABLED(LASER)
-        case PRINTER_MODE_LASER: {
-          if (laser.status != LASER_OFF) {
-            laser.status = LASER_OFF;
-            laser.mode = CONTINUOUS;
-            laser.duration = 0;
-
-            if (laser.diagnostics)
-              SERIAL_EM("Laser M5 called and laser OFF");
-          }
-        }
-        break;
-      #endif
-
-      #if ENABLED(CNCROUTER)
-        case PRINTER_MODE_CNC:
-          disable_cncrouter();
-        break;
-      #endif
-
-      default: break; // other tools
-
-    } // printer.mode
-
-    mechanics.prepare_move_to_destination();
-
-  }
-
-  #if ENABLED(CNCROUTER)
-    /*
-     * M6: CNC tool change
-     */
-    inline void gcode_M6() { tool_change_cnc(CNC_M6_TOOL_ID); }
-  #endif
-
-#endif // HAS_MULTI_MODE
-
-/**
- * M17: Enable power on all stepper motors
- */
-inline void gcode_M17() {
-  LCD_MESSAGEPGM(MSG_NO_MOVE);
-  stepper.enable_all_steppers();
-}
-
-#if HAS_SDSUPPORT
-
-  /**
-   * M20: List SD card to serial output
-   */
-  inline void gcode_M20() {
-    SERIAL_EM(MSG_BEGIN_FILE_LIST);
-    card.ls();
-    SERIAL_EM(MSG_END_FILE_LIST);
-  }
-
-  /**
-   * M21: Init SD Card
-   */
-  inline void gcode_M21() {
-    card.mount();
-  }
-
-  /**
-   * M22: Release SD Card
-   */
-  inline void gcode_M22() {
-    card.unmount();
-  }
-
-  /**
-   * M23: Select a file
-   */
-  inline void gcode_M23() {
-    // Simplify3D includes the size, so zero out all spaces (#7227)
-    for (char *fn = parser.string_arg; *fn; ++fn) if (*fn == ' ') *fn = '\0';
-    card.selectFile(parser.string_arg);
-  }
-
-  /**
-   * M24: Start or Resume SD Print
-   */
-  inline void gcode_M24() {
-    #if ENABLED(PARK_HEAD_ON_PAUSE)
-      printer.resume_print();
-    #endif
-
-    card.startFileprint();
-    printer.print_job_counter.start();
-    #if HAS_POWER_CONSUMPTION_SENSOR
-      startpower = power_consumption_hour;
-    #endif
-  }
-
-  /**
-   * M25: Pause SD Print
-   */
-  void gcode_M25() {
-    card.pauseSDPrint();
-    printer.print_job_counter.pause();
-    SERIAL_LM(REQUEST_PAUSE, "SD pause");
-
-    #if ENABLED(PARK_HEAD_ON_PAUSE)
-      commands.enqueue_and_echo_commands_P(PSTR("M125")); // Must be enqueued with pauseSDPrint set to be last in the buffer
-    #endif
-  }
-
-  /**
-   * M26: Set SD Card file index
-   */
-  inline void gcode_M26() {
-    if (card.cardOK && parser.seen('S'))
-      card.setIndex(parser.value_long());
-  }
-
-  /**
-   * M27: Get SD Card status
-   */
-  inline void gcode_M27() {
-    card.printStatus();
-  }
-
-  /**
-   * M28: Start SD Write
-   */
-  inline void gcode_M28() {
-    card.startWrite(parser.string_arg, false);
-  }
-
-  /**
-   * M29: Stop SD Write
-   * Processed in write to file routine above
-   */
-  inline void gcode_M29() {
-    // card.saving = false;
-  }
-
-  /**
-   * M30 <filename>: Delete SD Card file
-   */
-  inline void gcode_M30() {
-    if (card.cardOK) {
-      card.closeFile();
-      card.deleteFile(parser.string_arg);
-    }
-  }
-
-#endif // SDSUPPORT
-
-/**
- * M31: Get the time since the start of SD Print (or last M109)
- */
-inline void gcode_M31() {
-  char buffer[21];
-  duration_t elapsed = printer.print_job_counter.duration();
-  elapsed.toString(buffer);
-  lcd_setstatus(buffer);
-
-  SERIAL_LMT(ECHO, MSG_PRINT_TIME, buffer);
-}
-
-#if HAS_SDSUPPORT
-
-  /**
-   * M32: Make Directory
-   */
-  inline void gcode_M32() {
-    if (card.cardOK) {
-      card.makeDirectory(parser.string_arg);
-      card.mount();
-    }
-  }
-
-  /**
-   * M33: Close File and store location in restart.gcode
-   */
-  inline void gcode_M33() {
-    printer.stopSDPrint(true);
-  }
-
-  /**
-   * M34: Select file and start SD print
-   */
-  inline void gcode_M34() {
-    if (card.sdprinting)
-      stepper.synchronize();
-
-    if (card.cardOK) {
-      char* namestartpos = (strchr(parser.string_arg, '@'));
-      if (namestartpos == NULL) {
-        namestartpos = parser.string_arg ; // default name position
-      }
-      else
-        namestartpos++; // to skip the '@'
-
-      SERIAL_MV("Open file: ", namestartpos);
-      SERIAL_EM(" and start print.");
-      card.selectFile(namestartpos);
-      if (parser.seenval('S')) card.setIndex(parser.value_long());
-
-      mechanics.feedrate_mm_s       = 20.0; // 20 units/sec
-      mechanics.feedrate_percentage = 100;  // 100% mechanics.feedrate_mm_s
-      card.startFileprint();
-      printer.print_job_counter.start();
-      #if HAS_POWER_CONSUMPTION_SENSOR
-        startpower = power_consumption_hour;
-      #endif
-    }
-  }
-
-  #if ENABLED(NEXTION)
-    /**
-     * M35: Upload Firmware to Nextion from SD
-     */
-    inline void gcode_M35() {
-      UploadNewFirmware();
-    }
-  #endif
-
-#endif // SDSUPPORT
-
-/**
- * M42: Change pin status via GCode
- *
- *  P<pin>  Pin number (LED if omitted)
- *  S<byte> Pin status from 0 - 255
- */
-inline void gcode_M42() {
-  if (!parser.seenval('S')) return;
-  const byte pin_status = parser.value_byte();
-
-  const Pin pin_number = parser.intval('P', LED_PIN);
-  if (pin_number < 0) return;
-
-  if (printer.pin_is_protected(pin_number)) {
-    SERIAL_LM(ER, MSG_ERR_PROTECTED_PIN);
-    return;
-  }
-
-  HAL::pinMode(pin_number, OUTPUT);
-  HAL::digitalWrite(pin_number, pin_status);
-  HAL::analogWrite(pin_number, pin_status);
-
-  #if FAN_COUNT > 0
-    switch (pin_number) {
-      #if HAS_FAN0
-        case FAN_PIN: printer.fanSpeeds[0] = pin_status; break;
-      #endif
-      #if HAS_FAN1
-        case FAN1_PIN: printer.fanSpeeds[1] = pin_status; break;
-      #endif
-      #if HAS_FAN2
-        case FAN2_PIN: printer.fanSpeeds[2] = pin_status; break;
-      #endif
-      #if HAS_FAN3
-        case FAN3_PIN: printer.fanSpeeds[3] = pin_status; break;
-      #endif
-    }
-  #endif
-}
-
-#if ENABLED(PINS_DEBUGGING)
-
-  #include "utility/pinsdebug.h"
-
-  inline void toggle_pins() {
-    const bool  I_flag  = parser.boolval('I');
-    const int   repeat  = parser.intval('R', 1),
-                start   = parser.intval('S'),
-                end     = parser.intval('E', LAST_PIN - 1),
-                wait    = parser.intval('W', 500);
-
-    for (Pin pin = start; pin <= end; pin++) {
-
-      if (!I_flag && printer.pin_is_protected(pin)) {
-        SERIAL_MV("Sensitive Pin: ", pin);
-        SERIAL_EM(" untouched.");
-      }
-      else {
-        SERIAL_MV("Pulsing Pin: ", pin);
-        HAL::pinMode(pin, OUTPUT);
-        for (int16_t j = 0; j < repeat; j++) {
-          HAL::digitalWrite(pin, 0);
-          printer.safe_delay(wait);
-          HAL::digitalWrite(pin, 1);
-          printer.safe_delay(wait);
-          HAL::digitalWrite(pin, 0);
-          printer.safe_delay(wait);
-        }
-      }
-      SERIAL_EOL();
-    }
-    SERIAL_EM("Done");
-  } // toggle_pins
-
-  inline void servo_probe_test(){
-    #if !(NUM_SERVOS >= 1 && HAS_SERVO_0)
-      SERIAL_LM(ER, "SERVO not setup");
-    #elif !HAS_Z_SERVO_PROBE
-      SERIAL_LM(ER, "Z_ENDSTOP_SERVO_NR not setup");
-    #else
-
-      const uint8_t probe_index = parser.seen('P') ? parser.value_byte() : Z_ENDSTOP_SERVO_NR;
-
-      SERIAL_EM("Servo probe test");
-      SERIAL_EMV(".  Using index:  ", probe_index);
-      SERIAL_EMV(".  Deploy angle: ", probe.z_servo_angle[0]);
-      SERIAL_EMV(".  Stow angle:   ", probe.z_servo_angle[1]);
-
-      bool probe_inverting;
-
-      #if HAS_Z_PROBE_PIN
-
-        #define PROBE_TEST_PIN Z_PROBE_PIN
-
-        SERIAL_EMV("Probe uses Z_MIN_PROBE_PIN: ", PROBE_TEST_PIN);
-        SERIAL_EM(".  Uses Z_PROBE_ENDSTOP_INVERTING (ignores Z_MIN_ENDSTOP_INVERTING)");
-        SERIAL_MSG(".  Z_PROBE_ENDSTOP_INVERTING: ");
-
-        #if Z_PROBE_ENDSTOP_INVERTING
-          SERIAL_EM("true");
-        #else
-          SERIAL_EM("false");
-        #endif
-
-        probe_inverting = Z_PROBE_ENDSTOP_INVERTING;
-
-      #elif HAS_Z_MIN
-
-        #define PROBE_TEST_PIN Z_MIN_PIN
-
-        SERIAL_EMV("Probe uses Z_MIN pin: ", PROBE_TEST_PIN);
-        SERIAL_EM(".  Uses Z_MIN_ENDSTOP_INVERTING (ignores Z_PROBE_ENDSTOP_INVERTING)");
-        SERIAL_MSG(".  Z_MIN_ENDSTOP_INVERTING: ");
-
-        #if Z_MIN_ENDSTOP_INVERTING
-          SERIAL_EM("true");
-        #else
-          SERIAL_EM("false");
-        #endif
-
-        probe_inverting = Z_MIN_ENDSTOP_INVERTING;
-
-      #endif
-
-      SERIAL_EM("Deploy & stow 4 times");
-      SET_INPUT_PULLUP(PROBE_TEST_PIN);
-      bool deploy_state, stow_state;
-      for (uint8_t i = 0; i < 4; i++) {
-        servo[probe_index].move(probe.z_servo_angle[0]); // deploy
-        printer.safe_delay(500);
-        deploy_state = digitalRead(PROBE_TEST_PIN);
-        servo[probe_index].move(probe.z_servo_angle[1]); // stow
-        printer.safe_delay(500);
-        stow_state = digitalRead(PROBE_TEST_PIN);
-      }
-      if (probe_inverting != deploy_state) SERIAL_EM("WARNING - INVERTING setting probably backwards");
-
-      refresh_cmd_timeout();
-
-      if (deploy_state != stow_state) {
-        SERIAL_EM("BLTouch clone detected");
-        if (deploy_state) {
-          SERIAL_EM(".  DEPLOYED state: HIGH (logic 1)");
-          SERIAL_EM(".  STOWED (triggered) state: LOW (logic 0)");
-        }
-        else {
-          SERIAL_EM(".  DEPLOYED state: LOW (logic 0)");
-          SERIAL_EM(".  STOWED (triggered) state: HIGH (logic 1)");
-        }
-        #if ENABLED(BLTOUCH)
-          SERIAL_EM("ERROR: BLTOUCH enabled - set this device up as a Z Servo Probe with inverting as true.");
-        #endif
-
-      }
-      else {                                              // measure active signal length
-        servo[probe_index].move(probe.z_servo_angle[0]);  // deploy
-        printer.safe_delay(500);
-        SERIAL_EM("please trigger probe");
-        uint16_t probe_counter = 0;
-
-        // Allow 30 seconds max for operator to trigger probe
-        for (uint16_t j = 0; j < 500 * 30 && probe_counter == 0 ; j++) {
-
-          printer.safe_delay(2);
-
-          if (0 == j % (500 * 1)) // keep cmd_timeout happy
-            refresh_cmd_timeout();
-
-          if (deploy_state != digitalRead(PROBE_TEST_PIN)) { // probe triggered
-
-            for (probe_counter = 1; probe_counter < 50 && (deploy_state != digitalRead(PROBE_TEST_PIN)); probe_counter ++)
-              printer.safe_delay(2);
-
-            if (probe_counter == 50)
-              SERIAL_EM("Z Servo Probe detected");   // >= 100mS active time
-            else if (probe_counter >= 2 )
-              SERIAL_EMV("BLTouch compatible probe detected - pulse width (+/- 4mS): ", probe_counter * 2 );   // allow 4 - 100mS pulse
-            else
-              SERIAL_EM("noise detected - please re-run test");   // less than 2mS pulse
-
-            servo[probe_index].move(probe.z_servo_angle[1]); // stow
-
-          } // pulse detected
-
-        } // for loop waiting for trigger
-
-        if (probe_counter == 0) SERIAL_EM("trigger not detected");
-
-      } // measure active signal length
-
-    #endif
-
-  } // servo_probe_test
-
-  /**
-   * M43: Pin debug - report pin state, watch pins, toggle pins and servo probe test/report
-   *
-   *  M43         - report name and state of pin(s)
-   *                  P<pin>  Pin to read or watch. If omitted, reads all pins.
-   *                  I       Flag to ignore MK4duo's pin protection.
-   *
-   *  M43 W       - Watch pins -reporting changes- until reset, click, or M108.
-   *                  P<pin>  Pin to read or watch. If omitted, read/watch all pins.
-   *                  I       Flag to ignore MK4duo's pin protection.
-   *
-   *  M43 E<bool> - Enable / disable background endstop monitoring
-   *                  - Machine continues to operate
-   *                  - Reports changes to endstops
-   *                  - Toggles LED when an endstop changes
-   *                  - Can not reliably catch the 5mS pulse from BLTouch type probes
-   *
-   *  M43 T       - Toggle pin(s) and report which pin is being toggled
-   *                  S<pin>  - Start Pin number.   If not given, will default to 0
-   *                  L<pin>  - End Pin number.   If not given, will default to last pin defined for this board
-   *                  I       - Flag to ignore MK4duo's pin protection.   Use with caution!!!!
-   *                  R       - Repeat pulses on each pin this number of times before continueing to next pin
-   *                  W       - Wait time (in miliseconds) between pulses.  If not given will default to 500
-   *
-   *  M43 S       - Servo probe test
-   *                  P<index> - Probe index (optional - defaults to 0)
-   */
-  inline void gcode_M43() {
-
-    if (parser.seen('T')) {   // must be first ot else it's "S" and "E" parameters will execute endstop or servo test
-      toggle_pins();
-      return;
-    }
-
-    // Enable or disable endstop monitoring
-    if (parser.seen('E')) {
-      endstop_monitor_flag = parser.value_bool();
-      SERIAL_MSG("endstop monitor ");
-      SERIAL_TXT(endstop_monitor_flag ? "en" : "dis");
-      SERIAL_EM("abled");
-      return;
-    }
-
-    if (parser.seen('S')) {
-      servo_probe_test();
-      return;
-    }
-
-    // Get the range of pins to test or watch
-    const uint8_t first_pin = parser.seen('P') ? parser.value_byte() : 0,
-                  last_pin = parser.seen('P') ? first_pin : LAST_PIN - 1;
-
-    if (first_pin > last_pin) return;
-
-    const bool ignore_protection = parser.seen('I') && parser.value_bool();
-
-    // Watch until click, M108, or reset
-    if (parser.seen('W') && parser.value_bool()) { // watch digital pins
-      SERIAL_EM("Watching pins");
-      byte pin_state[last_pin - first_pin + 1];
-      for (int8_t pin = first_pin; pin <= last_pin; pin++) {
-        if (printer.pin_is_protected(pin) && !ignore_protection) continue;
-        pinMode(pin, INPUT_PULLUP);
-        // if (IS_ANALOG(pin))
-        //   pin_state[pin - first_pin] = analogRead(pin - analogInputToDigitalPin(0)); // int16_t pin_state[...]
-        // else
-          pin_state[pin - first_pin] = digitalRead(pin);
-      }
-
-      #if HAS(RESUME_CONTINUE)
-        wait_for_user = true;
-        KEEPALIVE_STATE(PAUSED_FOR_USER);
-      #endif
-
-      for(;;) {
-        for (int8_t pin = first_pin; pin <= last_pin; pin++) {
-          if (printer.pin_is_protected(pin)) continue;
-          byte val;
-          // if (IS_ANALOG(pin))
-          //   val = analogRead(pin - analogInputToDigitalPin(0)); // int16_t val
-          // else
-            val = digitalRead(pin);
-          if (val != pin_state[pin - first_pin]) {
-            report_pin_state(pin);
-            pin_state[pin - first_pin] = val;
-          }
-        }
-
-        #if HAS(RESUME_CONTINUE)
-          if (!wait_for_user) {
-            KEEPALIVE_STATE(IN_HANDLER);
-            break;
-          }
-        #endif
-
-        printer.safe_delay(500);
-      }
-      return;
-    }
-
-    // Report current state of selected pin(s)
-    for (uint8_t pin = first_pin; pin <= last_pin; pin++)
-      report_pin_state_extended(pin, ignore_protection);
-  }
-
-#endif // PINS_DEBUGGING
-
-#if ENABLED(Z_MIN_PROBE_REPEATABILITY_TEST)
-
-  /**
-   * M48: Z-Probe repeatability measurement function.
-   *
-   * Usage:
-   *   M48 <P#> <X#> <Y#> <V#> <E> <L#> <S>
-   *     P = Number of sampled points (4-50, default 10)
-   *     X = Sample X position
-   *     Y = Sample Y position
-   *     V = Verbose level (0-4, default=1)
-   *     E = Engage probe for each reading
-   *     L = Number of legs of movement before probe
-   *     S = Schizoid (Or Star if you prefer)
-   *
-   * This function assumes the bed has been homed.  Specifically, that a G28 command
-   * as been issued prior to invoking the M48 Z-Probe repeatability measurement function.
-   * Any information generated by a prior G29 Bed leveling command will be lost and need to be
-   * regenerated.
-   */
-  inline void gcode_M48() {
-
-    if (mechanics.axis_unhomed_error()) return;
-
-    int8_t verbose_level = parser.seen('V') ? parser.value_byte() : 1;
-    if (!WITHIN(verbose_level, 0, 4)) {
-      SERIAL_LM(ER, "?Verbose Level not plausible (0-4).");
-      return;
-    }
-
-    if (verbose_level > 0)
-      SERIAL_EM("M48 Z-Probe Repeatability Test");
-
-    int8_t n_samples = parser.seen('P') ? parser.value_byte() : 10;
-    if (!WITHIN(n_samples, 4, 50)) {
-      SERIAL_LM(ER, "?Sample size not plausible (4-50).");
-      return;
-    }
-
-    float X_current = mechanics.current_position[X_AXIS],
-          Y_current = mechanics.current_position[Y_AXIS];
-
-    bool stow_probe_after_each = parser.seen('E');
-
-    float X_probe_location = parser.seen('X') ? parser.value_linear_units() : X_current + X_PROBE_OFFSET_FROM_NOZZLE;
-    float Y_probe_location = parser.seen('Y') ? parser.value_linear_units() : Y_current + Y_PROBE_OFFSET_FROM_NOZZLE;
-
-    #if NOMECH(DELTA)
-      if (!WITHIN(X_probe_location, LOGICAL_X_POSITION(MIN_PROBE_X), LOGICAL_X_POSITION(MAX_PROBE_X))) {
-        out_of_range_error(PSTR("X"));
-        return;
-      }
-      if (!WITHIN(Y_probe_location, LOGICAL_Y_POSITION(MIN_PROBE_Y), LOGICAL_Y_POSITION(MAX_PROBE_Y))) {
-        out_of_range_error(PSTR("Y"));
-        return;
-      }
-    #else
-      if (!mechanics.position_is_reachable_by_probe_xy(X_probe_location, Y_probe_location)) {
-        SERIAL_LM(ER, "? (X,Y) location outside of probeable radius.");
-        return;
-      }
-    #endif
-
-    bool seen_L = parser.seen('L');
-    uint8_t n_legs = seen_L ? parser.value_byte() : 0;
-    if (n_legs > 15) {
-      SERIAL_LM(ER, "?Number of legs in movement not plausible (0-15).");
-      return;
-    }
-    if (n_legs == 1) n_legs = 2;
-
-    bool schizoid_flag = parser.seen('S');
-    if (schizoid_flag && !seen_L) n_legs = 7;
-
-    /**
-     * Now get everything to the specified probe point So we can safely do a
-     * probe to get us close to the bed.  If the Z-Axis is far from the bed,
-     * we don't want to use that as a starting point for each probe.
-     */
-    if (verbose_level > 2)
-      SERIAL_EM("Positioning the probe...");
-
-    // Disable bed level correction in M48 because we want the raw data when we probe
-    #if HAS_LEVELING
-      const bool was_enabled = bedlevel.leveling_is_active();
-      bedlevel.set_bed_leveling_enabled(false);
-    #endif
-
-    setup_for_endstop_or_probe_move();
-
-    // Move to the first point, deploy, and probe
-    const float t = probe.check_pt(X_probe_location, Y_probe_location, stow_probe_after_each, verbose_level);
-    if (isnan(t)) return;
-
-    randomSeed(millis());
-
-    double mean = 0.0, sigma = 0.0, min = 99999.9, max = -99999.9, sample_set[n_samples];
-
-    for (uint8_t n = 0; n < n_samples; n++) {
-      if (n_legs) {
-        int dir = (random(0, 10) > 5.0) ? -1 : 1;  // clockwise or counter clockwise
-        float angle = random(0.0, 360.0),
-              radius = random(
-                #if MECH(DELTA)
-                  mechanics.delta_probe_radius / 8, mechanics.delta_probe_radius / 3
-                #else
-                  5, X_MAX_LENGTH / 8
-                #endif
-              );
-
-        if (verbose_level > 3) {
-          SERIAL_MV("Starting radius: ", radius);
-          SERIAL_MV("   angle: ", angle);
-          SERIAL_MSG(" Direction: ");
-          if (dir > 0) SERIAL_MSG("Counter-");
-          SERIAL_EM("Clockwise");
-        }
-
-        for (uint8_t l = 0; l < n_legs - 1; l++) {
-          double delta_angle;
-
-          if (schizoid_flag)
-            // The points of a 5 point star are 72 degrees apart.  We need to
-            // skip a point and go to the next one on the star.
-            delta_angle = dir * 2.0 * 72.0;
-
-          else
-            // If we do this line, we are just trying to move further
-            // around the circle.
-            delta_angle = dir * (float) random(25, 45);
-
-          angle += delta_angle;
-
-          while (angle > 360.0)   // We probably do not need to keep the angle between 0 and 2*PI, but the
-            angle -= 360.0;       // Arduino documentation says the trig functions should not be given values
-          while (angle < 0.0)     // outside of this range.   It looks like they behave correctly with
-            angle += 360.0;       // numbers outside of the range, but just to be safe we clamp them.
-
-          X_current = X_probe_location - (X_PROBE_OFFSET_FROM_NOZZLE) + cos(RADIANS(angle)) * radius;
-          Y_current = Y_probe_location - (Y_PROBE_OFFSET_FROM_NOZZLE) + sin(RADIANS(angle)) * radius;
-
-          #if MECH(DELTA)
-            // If we have gone out too far, we can do a simple fix and scale the numbers
-            // back in closer to the origin.
-            while (!mechanics.position_is_reachable_by_probe_xy(X_current, Y_current)) {
-              X_current *= 0.8;
-              Y_current *= 0.8;
-              if (verbose_level > 3) {
-                SERIAL_MV("Pulling point towards center:", X_current);
-                SERIAL_EMV(", ", Y_current);
-              }
-            }
-          #else
-            X_current = constrain(X_current, X_MIN_POS, X_MAX_POS);
-            Y_current = constrain(Y_current, Y_MIN_POS, Y_MAX_POS);
-          #endif
-          if (verbose_level > 3) {
-            SERIAL_MSG("Going to:");
-            SERIAL_MV(" x: ", X_current);
-            SERIAL_MV(" y: ", Y_current);
-            SERIAL_EMV("  z: ", mechanics.current_position[Z_AXIS]);
-          }
-          mechanics.do_blocking_move_to_xy(X_current, Y_current);
-        } // n_legs loop
-      } // n_legs
-
-      // Probe a single point
-      sample_set[n] = probe.check_pt(X_probe_location, Y_probe_location, stow_probe_after_each, 0);
-
-      /**
-       * Get the current mean for the data points we have so far
-       */
-      double sum = 0.0;
-      for (uint8_t j = 0; j <= n; j++) sum += sample_set[j];
-      mean = sum / (n + 1);
-
-      NOMORE(min, sample_set[n]);
-      NOLESS(max, sample_set[n]);
-
-      /**
-       * Now, use that mean to calculate the standard deviation for the
-       * data points we have so far
-       */
-      sum = 0.0;
-      for (uint8_t j = 0; j <= n; j++)
-        sum += sq(sample_set[j] - mean);
-
-      sigma = SQRT(sum / (n + 1));
-      if (verbose_level > 0) {
-        if (verbose_level > 1) {
-          SERIAL_VAL(n + 1);
-          SERIAL_MV(" of ", (int)n_samples);
-          SERIAL_MV(": z: ", sample_set[n], 3);
-          if (verbose_level > 2) {
-            SERIAL_MV(" mean: ", mean, 4);
-            SERIAL_MV(" sigma: ", sigma, 6);
-            SERIAL_MV(" min: ", min, 3);
-            SERIAL_MV(" max: ", max, 3);
-            SERIAL_MV(" range: ", max - min, 3);
-          }
-          SERIAL_EOL();
-        }
-      }
-
-    }  // End of probe loop
-
-    if (probe.set_deployed(false)) return;
-
-    SERIAL_EM("Finished!");
-
-    if (verbose_level > 0) {
-      SERIAL_MV("Mean: ", mean, 6);
-      SERIAL_MV(" Min: ", min, 3);
-      SERIAL_MV(" Max: ", max, 3);
-      SERIAL_MV(" Range: ", max - min, 3);
-      SERIAL_EOL();
-    }
-
-    SERIAL_EMV("Standard Deviation: ", sigma, 6);
-    SERIAL_EOL();
-
-    clean_up_after_endstop_or_probe_move();
-
-    // Re-enable bed level correction if it had been on
-    #if HAS_LEVELING
-      bedlevel.set_bed_leveling_enabled(was_enabled);
-    #endif
-
-    mechanics.report_current_position();
-  }
-
-#endif // Z_MIN_PROBE_REPEATABILITY_TEST
-
-#if HAS_POWER_CONSUMPTION_SENSOR
-  /**
-   * M70 - Power consumption sensor calibration
-   *
-   * Z - Calibrate zero current offset
-   * A - Isert readed DC Current value (Ampere)
-   * W - Insert readed AC Wattage value (Watt)
-   */
-  inline void gcode_M70() {
-    if(parser.seen('Z')) {
-      SERIAL_EMV("Actual POWER_ZERO:", POWER_ZERO, 7);
-      SERIAL_EMV("New POWER_ZERO:", raw_analog2voltage(), 7);
-      SERIAL_EM("Insert new calculated values into the FW and call \"M70 A\" for the next calibration step.");
-    }
-    else if(parser.seen('A')) {
-      SERIAL_EMV("Actual POWER_ERROR:", POWER_ERROR, 7);
-      SERIAL_EMV("New POWER_ERROR:", analog2error(parser.value_float()), 7);
-      SERIAL_EM("Insert new calculated values into the FW and call \"M70 W\" for the last calibration step.");
-    }
-    else if(parser.seen('W')) {
-      SERIAL_EMV("Actual POWER_EFFICIENCY:", POWER_EFFICIENCY, 7);
-      SERIAL_EMV("New POWER_EFFICIENCY:", analog2efficiency(parser.value_float()), 7);
-      SERIAL_EM("Insert new calculated values into the FW and then ACS712 it should be calibrated correctly.");
-    }
-  }
-#endif
-
-/**
- * M75: Start print timer
- */
-inline void gcode_M75() { printer.print_job_counter.start(); }
-
-/**
- * M76: Pause print timer
- */
-inline void gcode_M76() { printer.print_job_counter.pause(); }
-
-/**
- * M77: Stop print timer
- */
-inline void gcode_M77() { printer.print_job_counter.stop(); }
-
-/**
- * M78: Show print statistics
- */
-inline void gcode_M78() {
-  // "M78 S78" will reset the statistics
-  if (parser.seen('S') && parser.value_int() == 78)
-    printer.print_job_counter.initStats();
-  else printer.print_job_counter.showStats();
-}
-
-#if HAS_POWER_SWITCH
-  /**
-   * M80   : Turn on Power Supply
-   * M80 S : Report the current state and exit
-   */
-  inline void gcode_M80() {
-
-    // S: Report the current power supply state and exit
-    if (parser.seen('S')) {
-      SERIAL_PS(powerManager.powersupply_on ? PSTR("PS:1\n") : PSTR("PS:0\n"));
-      return;
-    }
-
-    powerManager.power_on();
-
-    // If you have a switch on suicide pin, this is useful
-    // if you want to start another print with suicide feature after
-    // a print without suicide...
-    #if HAS(SUICIDE)
-      OUT_WRITE(SUICIDE_PIN, HIGH);
-    #endif
-
-    #if ENABLED(HAVE_TMC2130)
-      delay(100);
-      tmc2130_init(); // Settings only stick when the driver has power
-    #endif
-
-    LCD_MESSAGEPGM(WELCOME_MSG);
-
-    #if ENABLED(LASER) && ENABLED(LASER_PERIPHERALS)
-      laser.peripherals_on();
-      laser.wait_for_peripherals();
-    #endif
-  }
-#endif // HAS_POWER_SWITCH
-
-/**
- * M81: Turn off Power, including Power Supply, if there is one.
- *
- *      This code should ALWAYS be available for EMERGENCY SHUTDOWN!
- */
-inline void gcode_M81() {
-  thermalManager.disable_all_heaters();
-  thermalManager.disable_all_coolers();
-  stepper.finish_and_disable();
-
-  #if FAN_COUNT > 0
-    LOOP_FAN() printer.fanSpeeds[f] = 0;
-    #if ENABLED(PROBING_FANS_OFF)
-      printer.fans_paused = false;
-      ZERO(printer.paused_fanSpeeds);
-    #endif
-  #endif
-
-  #if ENABLED(LASER)
-    laser.extinguish();
-    #if ENABLED(LASER_PERIPHERALS)
-      laser.peripherals_off();
-    #endif
-  #endif
-
-  #if ENABLED(CNCROUTER)
-    disable_cncrouter();
-  #endif
-
-  printer.safe_delay(1000); // Wait 1 second before switching off
-
-  #if HAS(SUICIDE)
-    stepper.synchronize();
-    suicide();
-  #elif HAS_POWER_SWITCH
-    powerManager.power_off();
-  #endif
-
-  LCD_MESSAGEPGM(MACHINE_NAME " " MSG_OFF ".");
-
-}
 
 /**
  * M82: Set E codes absolute (default)
@@ -3187,7 +2183,7 @@ inline void gcode_M503() {
 
 #if HAS_EXT_ENCODER
   /**
-   * M512: Print Encoder status
+   * M512: Print Extruder Encoder status Pin
    */
   inline void gcode_M512() {
 
@@ -3352,7 +2348,7 @@ inline void gcode_M532() {
   inline void gcode_M600() {
 
     // Homing first
-    if (mechanics.axis_unhomed_error()) home_all_axes();
+    if (mechanics.axis_unhomed_error()) mechanics.Home(true);
 
     // Initial retract before move to pause park position
     const float retract = parser.seen('E') ? parser.value_axis_units(E_AXIS) : 0
@@ -3431,7 +2427,10 @@ inline void gcode_M532() {
    */
 
   /**
-   * M604: Set Extruder Encoder
+   * M604: Set data Extruder Encoder
+   *
+   *  S[step] - Set Error Steps
+   *
    */
   inline void gcode_M604() {
     GET_TARGET_EXTRUDER(604);
