+++ MK4duo/src/gcode/stats/m75.h
@@ -21,11 +21,14 @@
  */
 
 /**
- * gcode.h
+ * mcode
  *
  * Copyright (C) 2017 Alberto Cotronei @MagoKimbra
  */
 
-#define G28
+#define CODE_M75
 
-inline void gcode_G28(void) { mechanics.Home(false); }
+/**
+ * M75: Start print timer
+ */
+inline void gcode_M75(void) { printer.print_job_counter.start(); }
